Fix ansible_failed_task with inner always and outer rescue blocks v2
Rather then trying to iterate states to figure out whether the next task
will be in a rescue portion of the block and if so set
ansible_failed_task/ansible_failed_result variables, we could just look
if rescue portion of the block exists in any block from top down to the
most nested one. That way we do not have to handle various corner cases
where iterating to next task would end up in an always section but
rescue section would be in an outer block and we would fail to set the
variable.

Fixes #43191

ci_complete